feat: controller e templates de Venda com listagem, registro e modal de itens

+++ src/main/java/com/ppwii/supermercado/impl/CategoriaServiceImpl.java
@@ -2,6 +2,7 @@ package com.ppwii.supermercado.impl;
 
 import com.ppwii.supermercado.model.Categoria;
 import com.ppwii.supermercado.repository.CategoriaRepository;
+import com.ppwii.supermercado.repository.ProdutoRepository;
 import com.ppwii.supermercado.service.ICategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,9 @@ public class CategoriaServiceImpl implements ICategoriaService {
 
     @Autowired
     private CategoriaRepository categoriaRepository;
+
+    @Autowired
+    private ProdutoRepository produtoRepository;
 
     @Override
     public Integer saveCategoria(Categoria categoria) {
@@ -32,6 +36,10 @@ public class CategoriaServiceImpl implements ICategoriaService {
 
     @Override
     public void deleteCategoriaById(Integer id) {
+        Categoria categoria = getCategoriaById(id);
+        if (produtoRepository.existsByCategoria(categoria)) {
+            throw new RuntimeException("Não é possível excluir esta categoria pois existem produtos associados a ela.");
+        }
         categoriaRepository.deleteById(id);
     }
 }
